add recipe edit feature

+++ src/main/java/com/google/sps/servlets/RecipeDetailsServlet.java
@@ -74,8 +74,8 @@ public class RecipeDetailsServlet extends HttpServlet {
     );
     request.setAttribute("imgURL", recipeEntity.getProperty("imgURL"));
     request.setAttribute("difficulty", recipeEntity.getProperty("difficulty"));
-    request.setAttribute("hour", (int) recipeEntity.getProperty("prep_time")/60);
-    request.setAttribute("min", (int) recipeEntity.getProperty("prep_time")%60);
+    request.setAttribute("hour", (long) recipeEntity.getProperty("prep_time")/60);
+    request.setAttribute("min", (long) recipeEntity.getProperty("prep_time")%60);
     request.setAttribute(
       "ingredients",
       recipeEntity.getProperty("ingredients")
